Update toolbox.rmd
With the most recent version of USAboundaries the code chunk in lines 683-690 was failing so I updated it to work with the newest version, using geom_sf(). I also removed the head(c18df) function call, as with the update the c18df object is now 19 columns wide, which made the print aesthetically displeasing. In addition I changed the description (line 681) to accomodate this change and removed.