Fixing broken links in "Exercises" section
Graphics for (1) "London Cycle Hire Journeys" and (2) "In Climbing Income Ladder, Location Matters" were inaccessible.

(1) - Original link was broken. I've replaced for another one hosted in James's own blog but that is still online.
(2) - Original link is payawalled. I've replaced for a link pointing to a blog that has a post analysing this graphic. I've pointed directly to the graphic itself because pointing to the blogpost would spoil the exercise.